Apply customizations to release 1

+++ src/Presentation/Nop.Web/Models/Common/HeaderLinksModel.cs
@@ -9,7 +9,7 @@ namespace Nop.Web.Models.Common
         public string CustomerName { get; set; }
         
         public bool ShoppingCartEnabled { get; set; }
-        public int ShoppingCartItems { get; set; }
+        public decimal ShoppingCartItems { get; set; }
         
         public bool WishlistEnabled { get; set; }
         public int WishlistItems { get; set; }
